Renamed integration-test:junit -> integration-test:testutils, since this is really what this package contains, and there are no JUnit test specifics left.
Issue: #34
