+++ archunit-integration-test/src/test/java/com/tngtech/archunit/integration/ExamplesIntegrationTest.java
@@ -90,11 +90,11 @@ import com.tngtech.archunit.example.thirdparty.ThirdPartySubClassWithProblem;
 import com.tngtech.archunit.example.web.AnnotatedController;
 import com.tngtech.archunit.example.web.InheritedControllerImpl;
 import com.tngtech.archunit.exampletest.SecurityTest;
-import com.tngtech.archunit.junit.CyclicErrorMatcher;
-import com.tngtech.archunit.junit.ExpectedClass;
-import com.tngtech.archunit.junit.ExpectedMethod;
-import com.tngtech.archunit.junit.ExpectedTestFailures;
-import com.tngtech.archunit.junit.MessageAssertionChain;
+import com.tngtech.archunit.testutils.CyclicErrorMatcher;
+import com.tngtech.archunit.testutils.ExpectedClass;
+import com.tngtech.archunit.testutils.ExpectedMethod;
+import com.tngtech.archunit.testutils.ExpectedTestFailures;
+import com.tngtech.archunit.testutils.MessageAssertionChain;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
 
@@ -110,19 +110,19 @@ import static com.tngtech.archunit.example.controller.two.UseCaseTwoController.d
 import static com.tngtech.archunit.example.core.VeryCentralCore.DO_CORE_STUFF_METHOD_NAME;
 import static com.tngtech.archunit.example.persistence.layerviolation.DaoCallingService.violateLayerRules;
 import static com.tngtech.archunit.example.service.ServiceViolatingLayerRules.illegalAccessToController;
-import static com.tngtech.archunit.junit.CyclicErrorMatcher.cycle;
-import static com.tngtech.archunit.junit.ExpectedAccess.callFromConstructor;
-import static com.tngtech.archunit.junit.ExpectedAccess.callFromMethod;
-import static com.tngtech.archunit.junit.ExpectedAccess.callFromStaticInitializer;
-import static com.tngtech.archunit.junit.ExpectedDependency.constructor;
-import static com.tngtech.archunit.junit.ExpectedDependency.field;
-import static com.tngtech.archunit.junit.ExpectedDependency.inheritanceFrom;
-import static com.tngtech.archunit.junit.ExpectedDependency.method;
-import static com.tngtech.archunit.junit.ExpectedLocation.javaClass;
-import static com.tngtech.archunit.junit.ExpectedNaming.simpleNameOf;
-import static com.tngtech.archunit.junit.ExpectedViolation.clazz;
-import static com.tngtech.archunit.junit.ExpectedViolation.javaPackageOf;
-import static com.tngtech.archunit.junit.SliceDependencyErrorMatcher.sliceDependency;
+import static com.tngtech.archunit.testutils.CyclicErrorMatcher.cycle;
+import static com.tngtech.archunit.testutils.ExpectedAccess.callFromConstructor;
+import static com.tngtech.archunit.testutils.ExpectedAccess.callFromMethod;
+import static com.tngtech.archunit.testutils.ExpectedAccess.callFromStaticInitializer;
+import static com.tngtech.archunit.testutils.ExpectedDependency.constructor;
+import static com.tngtech.archunit.testutils.ExpectedDependency.field;
+import static com.tngtech.archunit.testutils.ExpectedDependency.inheritanceFrom;
+import static com.tngtech.archunit.testutils.ExpectedDependency.method;
+import static com.tngtech.archunit.testutils.ExpectedLocation.javaClass;
+import static com.tngtech.archunit.testutils.ExpectedNaming.simpleNameOf;
+import static com.tngtech.archunit.testutils.ExpectedViolation.clazz;
+import static com.tngtech.archunit.testutils.ExpectedViolation.javaPackageOf;
+import static com.tngtech.archunit.testutils.SliceDependencyErrorMatcher.sliceDependency;
 import static java.lang.System.lineSeparator;
 
 class ExamplesIntegrationTest {
