+++ archunit-integration-test/src/test/java/com/tngtech/archunit/testutils/ExpectedClass.java
@@ -1,4 +1,4 @@
-package com.tngtech.archunit.junit;
+package com.tngtech.archunit.testutils;
 
 public class ExpectedClass {
     public static Creator javaClass(Class<?> clazz) {
