+++ archunit-integration-test/src/test/java/com/tngtech/archunit/testutils/ExpectedNaming.java
@@ -1,4 +1,4 @@
-package com.tngtech.archunit.junit;
+package com.tngtech.archunit.testutils;
 
 public class ExpectedNaming {
     public static Creator simpleNameOf(Class<?> clazz) {
